requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.0
python-dotenv>=1.0
rich>=13.0
httpx>=0.27
numpy>=1.24
sentence-transformers>=2.2
faiss-cpu>=1.7
streamlit>=1.30
plotly>=5.0
rapidfuzz>=3.0
